@@ -4,14 +4,16 @@ version: 3.0.0
 homepage: https://github.com/gunschu/jitsi_meet
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  jitsi_meet_platform_interface: ^1.0.0
-  jitsi_meet_web_plugin: ^1.0.0
+  jitsi_meet_platform_interface: 
+    path: ../jitsi_meet_platform_interface
+  jitsi_meet_web_plugin: 
+    path: ../jitsi_meet_web_plugin
 
 dev_dependencies:
   flutter_test:
