@@ -4,22 +4,23 @@ version: 1.0.0
 homepage: https://github.com/gunschu/jitsi_meet
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.20.0"
   
 dependencies:
   flutter:
     sdk: flutter
-  js: ^0.6.2
-  meta: ^1.1.7
+  js: ^0.6.3
+  meta: ^1.3.0
   flutter_web_plugins:
     sdk: flutter
-  jitsi_meet_platform_interface: ^1.0.0
+  jitsi_meet_platform_interface: 
+    path: ../jitsi_meet_platform_interface
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.8.0
+  pedantic: ^1.11.0
   
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
