@@ -4,27 +4,30 @@ import 'feature_flag/feature_flag_enum.dart';
 import 'feature_flag/feature_flag_helper.dart';
 
 class JitsiMeetingOptions {
-  String room;
-  String serverURL;
-  String subject;
-  String token;
-  bool audioMuted;
-  bool audioOnly;
-  bool videoMuted;
-  String userDisplayName;
-  String userEmail;
-  String iosAppBarRGBAColor;
-  String userAvatarURL;
+  JitsiMeetingOptions({
+    required this.room,
+  });
 
-  Map<String, dynamic> webOptions; // options for web
+  final String room;
+  String? serverURL;
+  String? subject;
+  String? token;
+  bool? audioMuted;
+  bool? audioOnly;
+  bool? videoMuted;
+  String? userDisplayName;
+  String? userEmail;
+  String? iosAppBarRGBAColor;
+  String? userAvatarURL;
+
+  Map<String, dynamic>? webOptions; // options for web
 
   Map<FeatureFlagEnum, bool> featureFlags = HashMap();
 
-
   /// Get feature flags Map with keys as String instead of Enum
   /// Useful as an argument sent to the Kotlin/Swift code
-  Map<String, bool> getFeatureFlags() {
-    Map<String, bool> featureFlagsWithStrings = HashMap();
+  Map<String?, bool> getFeatureFlags() {
+    Map<String?, bool> featureFlagsWithStrings = HashMap();
 
     featureFlags.forEach((key, value) {
       featureFlagsWithStrings[FeatureFlagHelper.featureFlags[key]] = value;
